--- modules/appeals_api/spec/requests/v1/decision_reviews/notice_of_disagreements/contestable_issues_spec.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-require 'rails_helper'
-require_relative AppealsApi::Engine.root.join('spec', 'support', 'shared_examples_contestable_issues.rb')
-
-Rspec.describe 'AppealsApi::V1::DecisionReviews::NoticeOfDisagreements::ContestableIssues', type: :request do
-  include_examples 'Appealable Issues API v0 and Decision Reviews v1 & v2 shared request examples',
-                   base_path: '/services/appeals/v1/decision_reviews/notice_of_disagreements/contestable_issues'
-end
